Commit 1 RankenUnofficial
I've been working on this project for a few weeks now. I've decided to upload this project to github, so I can easily work on it from both my Desktop and my Laptop. It's still a work in progress. However I'm proud of the work I've done so far. I will make a ToDo list ASAP so I can have a clear road map.